+++ printing/lib/src/preview/pdf_preview.dart
@@ -346,11 +346,11 @@ class _PdfPreviewState extends State<PdfPreview> {
       actions.add(PdfPageFormatAction(
         pageFormats: widget.pageFormats,
       ));
+    }
 
-      if (widget.useActions && widget.canChangeOrientation) {
-        // ignore: prefer_const_constructors
-        actions.add(PdfPageOrientationAction());
-      }
+    if (widget.useActions && widget.canChangeOrientation) {
+      // ignore: prefer_const_constructors
+      actions.add(PdfPageOrientationAction());
     }
 
     widget.actions?.forEach(actions.add);